better code :)

+++ dht11.h
@@ -11,8 +11,9 @@ struct DHTobj
     uint8_t sensor;
     bool ack;
 };
-bool read_dht11(int* dht11_dat);
-bool dht11_read(int* dht11_val);
+
+bool read_dht11(uint8_t* dht11_dat);
+bool dht11_read(uint8_t* dht11_val);
 bool dht11_read_obj(DHTobj* DHTdata);
 
 #endif
